requirements.txt
discord-rewrite==1.0.0
yarl<1.2